requirements.txt
Django==3.1
drf-spectacular
djangorestframework==3.11
psycopg2-binary>=2.8
django-cors-headers
django-filter
django-crispy-forms
Faker
django-concurrency
django-celery-beat
celery
coverage